refactor: move pool helper to utils
